Move open_session
