+++ omc4py/__init__.py
@@ -32,55 +32,4 @@ __license__ = '''
  */
 '''
 
-import typing
-
-from . import (
-    classes,
-    compiler,
-    session,
-)
-
-
-def open_session(
-    omc_command: typing.Optional[compiler.StrOrPathLike] = None,
-    *,
-    session_type: typing.Optional[
-        typing.Type[classes.AbstractOMCSession]
-    ] = None,
-) -> classes.AbstractOMCSession:
-    omc = compiler.OMCInteractive.open(omc_command)
-
-    try:
-        if session_type is None:
-            session_type = __select_session_type(omc)
-        else:
-            if not issubclass(session_type, classes.AbstractOMCSession):
-                raise TypeError(
-                    "session_type must be "
-                    f"subclass of {classes.AbstractOMCSession}, "
-                    f"got {session_type}"
-                )
-
-        return session_type(omc)
-
-    except Exception:
-        omc.close()
-        raise
-
-
-def __select_session_type(
-    omc: classes.AbstractOMCInteractive,
-) -> typing.Type[classes.AbstractOMCSession]:
-    version = session.OMCSessionMinimal(omc).getVersionTuple()
-    if version[:2] <= (1, 13):
-        from . import v_1_13
-        return v_1_13.OMCSession
-    elif version[:2] == (1, 14):
-        from . import v_1_14
-        return v_1_14.OMCSession
-    elif version[:2] == (1, 15):
-        from . import v_1_15
-        return v_1_15.OMCSession
-    else:  # version[:2] >= (1, 16):
-        from . import v_1_16
-        return v_1_16.OMCSession
+from .session import open_session
